LayoutTests/http/tests/security/contentSecurityPolicy/script-src-self-blocked-03-expected.txt

CONSOLE MESSAGE: Refused to load https://127.0.0.1:8443/security/contentSecurityPolicy/resources/script.js because it does not appear in the script-src directive of the Content Security Policy.


--------
Frame: '<!--framePath //<!--frame0-->-->'
--------
PASS
